指摘箇所の修正

+++ src/Review01.java
@@ -3,13 +3,13 @@ public class Review01 {
 
     public static void main(String[] args) {
         int p = 1500;
-        double tax = tax(p);
-        System.out.println(p + "円の商品の税込価格は" + (int)(p + tax) + "円（消費税は" + (int)tax + "円）です。");
+        int tax = tax(p);
+        System.out.println(p + "円の商品の税込価格は" + (p + tax) + "円（消費税は" + tax + "円）です。");
 
     }
 
-    public static double tax(int p) {
-        double tax = p * 0.1;
+    public static int tax(int p) {
+        int tax = (int)(p * 0.1);
         return tax;
     }
